test(mail): move cross-crate coverage
