+++ tests/integration/tests/mail.rs
@@ -1,0 +1,7 @@
+// Integration tests for settings-driven mail backends and role-mail helpers.
+
+#[path = "mail/backend_settings_integration.rs"]
+mod backend_settings_integration;
+
+#[path = "mail/role_mail_integration.rs"]
+mod role_mail_integration;
